requirements.txt
pymongo==3.10.1
Flask==1.1.1
PyJWT==1.7.1
